Have created "help.ejs" file and separate CSS file for "index.ejs"
The "style" CSS file for the landing page has been placed into a folder named "index" under the main CSS folder. This is because the landing page (index.ejs) will have it's own style of background. Have also created "help.ejs" file which will be the F&Q page for the web app. Deleted (test.ejs) web page as this was used as test.