Have included arrow SVG icons within each question button.
Have included the arrow SVGs that will point downwards from it's right pointing position when the collapsible question is clicked on to reveal the answer. Have styled the svg arrows with the "arrow-icon" CSS class. Have created the "answer-content" class that will appear once a question button has been clicked on revealing the answer text.