Have included CSS border properties for "wrapper" class.
Created hover colour effect for "drop-down" button class used in the "help.ejs" web page. The ".wrapper .active" class toggles the set border-radius for collapsible section for a question on the help page. The "arrow-icon.active" class has been given a transform rotation property so that the arrow svg icon rotates at 90 degrees to the right when toggle active for the question/answer collapsible feature.