Have included two small cards in Landing Page "index.ejs" page.
The left small card will store a welcome message for the logged in user as well as a button directing them to the to the To-Do list page. The right small card will store a time and date greeting message showing the time, date and a message saying "good morning, good afternoon or good evening". Have made "page-button" class and have styled card text and positioned items to centre of each card.