Have included finishing front end features to Index page.
Have included two small cards that have been styled. The left small card will contain a welcoming message to the logged in user with a placeholder value of text being changed showing their account name. This will be after the text characters that quotes "Welcome:". A button has also been placed beneath which links to the to-do-list web page. The right small card contains a dynamic message which will greet the user with a morning, afternoon or evening message depending on the time of day. The time will also be shown in a 24 hour digital clock format. Functionality made with JS. The three span id's have been made for this.