Functionality coded for time greet message and time display for index page.
Coded functionality in "index-page.js" for index page. "timeGreeting" function containing if statement for displaying greeting message depending on the time of day using JS Date.getHours object. For example "Hello there & Good Evening!" The "timeMessage", "showHour" and "showMinute" are linked to the HTML span tag id's that will display the correct and corresponding text in HTML. These are what show the greeting message, hour time and minute time of that hour.