Have added the div class "list wrapper" to the HTML body.
Have added the to-do list HTML skeleton within a wrapper div class called "list-warpper". This includes a form that will use a POST method with an input box for text and button to submit and create a task. An unordered list has been used storing one item. This will act as displaying a task card or item as they will be made dynamically when text is typed in and added within the input box.