Have implemented more constants in the index.js file.
"const TodoTask = require("./models/TodoTask");" is to have access to the TodoTask data model JS file for the db.

Have instantiation the todoTask const data model object in the app.post method to allow body content of the to-do-list web page to be requested with the data.

The app.get method for rendering the to-do-list page contains "res.render('to-do-list.ejs', { todoTask: tasks});" which is the name of the data model for holding the task cards in the mongo db schema.