Have added node templating into the HTML for the "to-do-list.ejs" page.
Using <% todoTask.forEach(details => { %>  <% }) %>

This enables CRUD Read functionality so data can be read from the mongo db collection named "todotasks" and can appear as a task card on the web page. When a new to-do list task card is created on the web page it will appear and will be stored in the database.