Have made some updates to api options in index.js file.
Have removed the app.get res render methods for the "sign-up" & "login" page as they are no longer needed in the project. Have made the index page route "/" so it acts as the main index/gateway page on localhost:3000

Had also changed the TCP host port for the web app from 8080 to 3000.