Changed the stylesheet css links by removing dot.
Removed dot at the start of the href links for the relevant CSS stylesheets. This is so the static accessor in the expressJS node api can access the stylesheets using app.use("/assets", express.static("assets"));. The to-do-list-edit.ejs was having problems rendering the CSS because of this so had to make this change.