Have placed the edit and delete svg icons in position.
Have placed the edit and delete svg icons into the edit and delete hyperlinks for the task cards on the to-do-list page. They have been coloured and styled in the style.css file for the to-do-list web page. Used the text "edit" & "delete" beforehand during the back-end development. Have now made the finishing touches to styling.