Have made the CRUD update and delete data methods.
These are the app route methods that deal with the update and delete of data for the tasks on the to-do-list web page. The TodoTask, "findByIdAndUpdate" & "findByIdAndDelete" objects are used to grab an id for a specific task from the mongo database. This is so a task can be updated with it's filled body text or simply deleted.