Have set up users web socket array in index.js
This is the io.on connection function that emits the broadcasting for the username and user connecting to the session, a message being sent from that user/client and the client disconnecting from the session.

Have also applied express api port number const to top of file instead of  storing it in the mongoose.connect function.